Add basic logged-in guard
For the ease of development it's use is commented out.
The added guard can be used if a client is logged in.
If they are not then they will be redirected to the login page.
A private boolean and a getter method were introduced to the
authorization service to make checking if a user is logged in
simpler and synchronous. The functionality of the getter method can be
extended in the future.